+++ quantify_scheduler/backends/qblox/qblox_hardware_config_old_style.py
@@ -57,6 +57,9 @@ hardware_config = {
                     {"port": "q6:mw", "clock": "q6.01", "interm_freq": None}
                 ],
             },
+            "digital_output_0": {
+                "portclock_configs": [{"port": "q0:switch", "clock": "digital"}]
+            },
         },
         "cluster0_module3": {
             "instrument_type": "QRM",